chore: delete files that are redundant
